backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
gevent==24.2.1
gevent-websocket==0.10.1
requests==2.31.0
python-dotenv==1.0.0

